feat: v2.2 architecture, retrieval, and UX phases + critical faithfulness fix
Phase 1 (architecture): split api_server.py (1507->~115 lines) into
routes/{query,chat,ingest,agent,management,feedback}.py with shared
deps.py; extract llm_client.py from rag.py; SQLite session/job
persistence (persistence.py) with WAL mode; request-ID middleware;
graceful shutdown via direct uvicorn.run().

Phase 2 (retrieval): ColBERT MaxSim reranking (colbert_rerank.py,
opt-in via USE_COLBERT_RERANK), HyDE retrieval (USE_HYDE), arXiv
metadata enrichment (metadata_enrich.py), cross-ingestion title dedup
(vector_store.find_similar_paper).

Phase 3 (UX): query_id + feedback endpoint/buttons, BibTeX export
(/search/export), opt-in user prefs storage (/prefs/{user_id}).

Bug fixes found via live agent testing:
- agent/tool_executor.py: query-expansion merge was truncating by
  append-order instead of re-ranking against the original query,
  dropping relevant papers from the final answer.
- pdf_utils.py: title extraction scored individual PyMuPDF spans
  instead of full lines/blocks, producing single-word garbage titles.
- verify.py: [Cite:N] markers were reaching the NLI model as part of
  the hypothesis text, collapsing entailment probability toward 0
  regardless of actual support (verified 0.998 -> 0.009 on an
  identical claim/chunk pair) — this was the root cause of
  faithfulness scores reading ~0 on nearly every real answer.
- persistence.py: job eviction only cleared the in-memory dict, never
  the SQLite row, so completed jobs accumulated on disk forever.
- tests/conftest.py: isolate test runs from the real sessions.db.
