+++ rag.py
@@ -9,43 +9,10 @@ import embeddings
 import vector_store
 import lang_utils
 import translation
-from google import genai
+import llm_client
 from google.genai import types
 
 logger = logging.getLogger(__name__)
-
-# Client pool — one genai.Client per API key, round-robin load balanced.
-# Lazily initialised so the module can be imported without keys (retrieval-only mode).
-_client_pool: list[genai.Client] = []
-_client_lock = __import__('threading').Lock()
-_client_index = __import__('itertools').cycle([])  # replaced on init
-
-
-def _init_client_pool() -> None:
-    """Build the client pool from config.LLM_API_KEY_POOL (called under lock)."""
-    global _client_pool, _client_index
-    if not config.LLM_API_KEY_POOL:
-        raise ValueError(
-            "Google Gemini API key not configured. "
-            "Set LLM_API_KEY (single) or LLM_API_KEYS (comma-separated) in .env."
-        )
-    _client_pool = [genai.Client(api_key=k) for k in config.LLM_API_KEY_POOL]
-    _client_index = __import__('itertools').cycle(range(len(_client_pool)))
-
-
-def _next_client_idx() -> int:
-    """Advance the round-robin counter under the pool lock (BUG-001/002)."""
-    with _client_lock:
-        return next(_client_index)
-
-
-def _get_client() -> genai.Client:
-    """Return the next client from the round-robin pool (thread-safe)."""
-    if not _client_pool:
-        with _client_lock:
-            if not _client_pool:
-                _init_client_pool()
-    return _client_pool[_next_client_idx()]
 
 
 def extract_citations(answer: str, metadatas: List[Dict], chunks: List[str] = None) -> List[Dict]:
@@ -92,11 +59,39 @@ def extract_citations(answer: str, metadatas: List[Dict], chunks: List[str] = No
     return citations
 
 
+def _hyde_embedding(user_query: str):
+    """Draft a hypothetical answer and embed it, for HyDE retrieval.
+
+    Bridges the lexical gap for complex/multi-hop queries: the hypothetical
+    answer's vocabulary overlaps documents more than the bare question does.
+    Falls back to embedding the raw query on any LLM failure.
+    """
+    try:
+        hyde_config = types.GenerateContentConfig(
+            temperature=config.LLM_TEMPERATURE,
+            max_output_tokens=256,
+            safety_settings=config.SAFETY_SETTINGS,
+        )
+        response = llm_client.generate_with_failover(
+            config.LLM_MODEL_NAME,
+            f"Write a short, plausible-sounding answer to this question, "
+            f"even if you are not sure it is correct:\n\n{user_query}",
+            hyde_config,
+        )
+        hypothetical = safe_extract_text(response)
+        if hypothetical:
+            return embeddings.embed_query(hypothetical)
+    except Exception as e:
+        logger.debug(f"HyDE draft failed, falling back to direct query embedding: {e}")
+    return embeddings.embed_query(user_query)
+
+
 def retrieve_context(
     user_query: str,
     top_k: int = None,
     filter_dict: Optional[Dict[str, Any]] = None,
-    collection=None
+    collection=None,
+    use_hyde: bool = None,
 ) -> Dict[str, Any]:
     """
     Retrieve relevant context for a user query.
@@ -117,11 +112,13 @@ def retrieve_context(
     """
     if top_k is None:
         top_k = config.DEFAULT_TOP_K
+    if use_hyde is None:
+        use_hyde = config.USE_HYDE
 
     from cache import retrieval_cache, make_key
     cache_scope = None if collection is None else getattr(collection, "name", id(collection))
     cache_key = make_key(user_query, top_k, filter_dict, cache_scope,
-                         config.USE_RERANKER, config.MAX_CONTEXT_CHUNKS)
+                         config.USE_RERANKER, config.MAX_CONTEXT_CHUNKS, use_hyde)
     if collection is None and filter_dict is None:
         cached = retrieval_cache.get(cache_key)
         if cached is not None:
@@ -131,8 +128,9 @@ def retrieve_context(
     if collection is None:
         collection = vector_store.get_or_create_collection()
 
-    # Embed the query
-    query_embedding = embeddings.embed_query(user_query)
+    # Embed the query — HyDE embeds a drafted hypothetical answer instead of
+    # the bare question; lexical (BM25) search below always uses the real query.
+    query_embedding = _hyde_embedding(user_query) if use_hyde else embeddings.embed_query(user_query)
     
     # Search vector store (dense)
     results = vector_store.search(
@@ -184,6 +182,14 @@ def retrieve_context(
     docs = results['documents']
     metas = results['metadatas']
     dists = results['distances']
+
+    if config.USE_COLBERT_RERANK and docs:
+        import colbert_rerank
+        dense_sims = [1.0 - d for d in dists]  # cosine distance -> similarity
+        colbert_top_k = min(len(docs), config.MAX_CONTEXT_CHUNKS * 3)
+        docs, metas, dists = colbert_rerank.rerank(
+            user_query, docs, metas, dense_sims,
+            top_k=colbert_top_k, weight=config.COLBERT_WEIGHT)
 
     if config.USE_RERANKER and docs:
         import rerank
@@ -315,7 +321,7 @@ def llm_generate(prompt: str, max_tokens: int = None,
     )
 
     try:
-        response = generate_with_failover(config.LLM_MODEL_NAME, prompt, generate_config)
+        response = llm_client.generate_with_failover(config.LLM_MODEL_NAME, prompt, generate_config)
 
         # Check if response has text
         if response.text:
@@ -351,40 +357,7 @@ def llm_generate(prompt: str, max_tokens: int = None,
         raise
 
 
-def llm_generate_stream(prompt: str, max_tokens: int = None, system_instruction: str = None):
-    """Generator: stream LLM response chunks using primary model with key cycling.
-
-    Yields non-empty text chunks as they arrive. No model failover — primary model only.
-    ponytail: single-model streaming; add failover if primary is unreliable.
-    """
-    if max_tokens is None:
-        max_tokens = config.LLM_MAX_TOKENS
-
-    if not _client_pool:
-        with _client_lock:
-            if not _client_pool:
-                _init_client_pool()
-
-    gen_config = types.GenerateContentConfig(
-        temperature=config.LLM_TEMPERATURE,
-        max_output_tokens=max_tokens,
-        safety_settings=config.SAFETY_SETTINGS,
-        system_instruction=system_instruction or config.SYSTEM_PROMPT,
-    )
-
-    client = _get_client()
-    emitted = False
-    for chunk in client.models.generate_content_stream(
-        model=config.LLM_MODEL_NAME, contents=prompt, config=gen_config
-    ):
-        try:
-            if chunk.text:
-                emitted = True
-                yield chunk.text
-        except (ValueError, AttributeError) as exc:
-            logger.debug("Skipping non-text Gemini stream chunk: %s", exc)
-    if not emitted:
-        raise RuntimeError("No text generated from Gemini stream")
+llm_generate_stream = llm_client.llm_generate_stream
 
 
 def prepare_query_for_stream(user_query: str, strategy: str = "A", top_k: int = None) -> dict:
@@ -482,79 +455,7 @@ def prepare_chat_for_stream(messages: List[Dict[str, str]], strategy: str = "A",
             "metadatas": context_data["metadatas"], "detected_lang": detected_lang, "lang_name": lang_name}
 
 
-def _is_transient(exc: Exception) -> bool:
-    status = getattr(exc, "status_code", None) or getattr(exc, "code", None)
-    if status in (429, 503):
-        return True
-    msg = str(exc)
-    return "503" in msg or "429" in msg or "UNAVAILABLE" in msg or "RESOURCE_EXHAUSTED" in msg
-
-
-import time as _time
-
-_circuit_breaker: dict[str, float] = {}
-_CIRCUIT_COOLDOWN = 60
-
-
-def generate_with_failover(model: str, contents, gen_config):
-    """
-    Call generate_content rotating through all API keys on 503/429 errors.
-
-    Tries every client in the pool exactly once.  If all keys fail on the
-    primary model, retries with LLM_FALLBACK_MODEL before giving up.
-    Uses a circuit breaker to skip models that recently failed on all keys.
-    """
-    if not _client_pool:
-        with _client_lock:
-            if not _client_pool:
-                _init_client_pool()
-
-    pool = _client_pool
-    models_to_try = [model]
-    if config.LLM_FALLBACK_MODEL and config.LLM_FALLBACK_MODEL != model:
-        models_to_try.append(config.LLM_FALLBACK_MODEL)
-
-    # Round-robin: start from the next key in rotation, not always pool[0]
-    start = _next_client_idx()
-    ordered_pool = pool[start:] + pool[:start]
-
-    last_exc: Exception | None = None
-    any_attempted = False
-    for current_model in models_to_try:
-        tripped_until = _circuit_breaker.get(current_model, 0)
-        if _time.monotonic() < tripped_until:
-            logger.info(f"[Gemini failover] {current_model} circuit open, skipping")
-            continue
-
-        any_attempted = True
-        all_failed = True
-        for offset, client in enumerate(ordered_pool, 1):
-            try:
-                result = client.models.generate_content(
-                    model=current_model, contents=contents, config=gen_config
-                )
-                _circuit_breaker.pop(current_model, None)
-                return result
-            except Exception as exc:
-                if _is_transient(exc):
-                    logger.warning(
-                        f"[Gemini failover] {current_model} key #{offset}/{len(pool)} "
-                        f"returned {getattr(exc, 'status_code', '?')}: {exc!s:.120} — trying next"
-                    )
-                    last_exc = exc
-                    continue
-                raise
-
-        if all_failed:
-            _circuit_breaker[current_model] = _time.monotonic() + _CIRCUIT_COOLDOWN
-            if current_model == model and len(models_to_try) > 1:
-                logger.warning(f"[Gemini failover] {model} circuit tripped for {_CIRCUIT_COOLDOWN}s, falling back to {config.LLM_FALLBACK_MODEL}")
-
-    if not any_attempted:
-        raise RuntimeError(
-            "All configured Gemini models are currently circuit-open; retry after cooldown."
-        )
-    raise last_exc  # type: ignore[misc]
+generate_with_failover = llm_client.generate_with_failover
 
 
 def safe_extract_text(response) -> str:
