+++ requirements.txt
@@ -5,6 +5,7 @@ torch>=2.12.0
 transformers>=4.57.6,<5.0.0    # hold below 5.x — major API rewrite
 safetensors>=0.8.0
 tokenizers>=0.22.2
+FlagEmbedding>=1.3.0           # BGE-M3 ColBERT multi-vector output (USE_COLBERT_RERANK)
 
 # PDF processing
 pymupdf>=1.27.0
